Fix deadly bugs, Finish exercice 2.91
The division operation on the scheme number was a remainder, a thing i
did because i thought i was smart and that operations concerning
integers should always return an integer.

Fix bug where the-empty-list was actually a function, not a value in
itself, this was infuriating and i don't know why they did it this way
...

Div-polys now returns the variable with the term list, not just the term
list.
i
OH ALSO, fixed an extremely deadly bug in mul terms, where i was
referencing an unidentified 'variable' ... the kind of bugs that make
you think ...