HAH, silly me.
The fact was that the folder defining the routines for LetStar was never even reached.
Because while loading EvalLet, it goes into the driver loop, cuz i forgot to uncomment (driver-loop) in EvalLet.
Shouldve realized it sooner when i noticed the incongruent fact that the (let*->nested-lets foo) didnt run, but oh well