Dear future me,
This is taking me a decent chunk of time.

Mostly because I am confused as to many things.

Even still, I hope you find this instructive and useful, do not let my
effort go to waste, learn from it.

LEARN FROM IT I SAY !!!11!!1!

In particular, analyze the parens of the result, adjust the code so it
generates the correct output, then conclude what and when (list ...) is
needed, when it is not, and all the details of whether a resulting
nested list need be quoted as is or if you need to map something along
the result, and so on ..

And don't worry about it, there are worse things you can be wasting your
time on.