Finally.
My GOD this took me a heccing while.
Well, that's just another hiccup along the way, hopefully.
Note that my implementation isn't perfect at all, but at this point as long as it works.
Gotta focus on the core of ideas, I reckon (altho there is room for improvement, but that will come later)