So, here's the plan.
First of all, added some vim folds in here.
Second of all, this thing ALyssa asked for stumped me more than I'd like
to admit.
Though I have a tentative plan, that is kinda probably sub-optimal, but
I'll try nevertheless.
I plan on giving each instruction a number, an order, I figured I can't
place it on extract-labels, so I am making the instruction making
procedure a procedure with local state (ugh) ...
And then probably when executing an instruction I will see to check if
the labels associated with it are to be founf in the breakpoints ... I
think? whi